Added Ack KnowhereMessage flow
Removed the old KnowhereDataMessage and replaced it with KnowhereMessage which itself is data from the KnowhereMessage table.
Added the ack flow so the add flow is completed. This means that after an add KnowhereMessage is sent from Device A to Device B,
the Device B will now add the message to the table then execute the add operation then generate an ack message with the remote_id
field and original message_id field set in the ack message and send it back to Device A.
Upon receiving the ack message Device A looks up the original add message and find the originally added target and update the remote_id field.
The flow of adding a target is then complete.