Fix gcm send message error over cellular radio
Gcm was not able to send message over radio only wifi and it turns out to be related the the api id.
In order to send gcm message over cellular radio it needs to use api id for browser application instead of
server application.
Also fixed the settings send message debug menu use for testing.